Fork and publish the gem under carwow's domain.
At carwow, we are committed to the open source community and our contributions will be published to the original gem and main repository as usual, but in the meantime we need to publish a new gem to carry on with our work.
We are not the official maintainers for `json_api_client` gem, but we are happy to contribute to the OS community with this fork.